fix warning code

+++ Category/NSString/NSString+Wrapper.m
@@ -59,7 +59,7 @@
 }
 
 - (int) indexOfChar:(unichar)ch fromIndex:(int)index{
-    int len = self.length;
+    int len = (int)self.length;
     for (int i = index; i < len; ++i) {
         if (ch == [self charAt:i]) {
             return i;
@@ -73,7 +73,7 @@
     if (range.location == NSNotFound) {
         return JavaNotFound;
     }
-    return range.location;
+    return (int)range.location;
 }
 
 - (int) indexOfString:(NSString*)str fromIndex:(int)index {
@@ -82,11 +82,11 @@
     if (range.location == NSNotFound) {
         return JavaNotFound;
     }
-    return range.location;
+    return (int)range.location;
 }
 
 - (int) lastIndexOfChar:(unichar)ch {
-    int len = self.length;
+    int len = (int)self.length;
     for (int i = len-1; i >=0; --i) {
         if ([self charAt:i] == ch) {
             return i;
@@ -96,7 +96,7 @@
 }
 
 - (int) lastIndexOfChar:(unichar)ch fromIndex:(int)index {
-    int len = self.length;
+    int len = (int)self.length;
     if (index >= len) {
         index = len - 1;
     }
@@ -113,7 +113,7 @@
     if (range.location == NSNotFound) {
         return JavaNotFound;
     }
-    return range.location;
+    return (int)range.location;
 }
 
 - (int) lastIndexOfString:(NSString*)str fromIndex:(int)index {
@@ -122,7 +122,7 @@
     if (range.location == NSNotFound) {
         return JavaNotFound;
     }
-    return range.location;
+    return (int)range.location;
 }
 
 - (NSString *) substringFromIndex:(int)beginIndex toIndex:(int)endIndex {
